server/flask/requirements.txt
flask
pymongo
selenium
flask-sqlalchemy
pymysql
uwsgi